# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/auth/revoke-all": {
      "post": {
        "operationId": "AdminAuthController_revokeAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRevokedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke every active session for the current Admin",
        "tags": [
          "Admin Auth"
        ]
      }
    },
    "/api/v1/admin/auth/me": {
      "post": {
        "operationId": "AdminAuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Return server-authoritative Admin identity and capabilities",
        "tags": [
          "Admin Auth"
        ]
      }
    },
...
    "/api/v1/member/auth/otp/request": {
      "post": {
        "operationId": "MemberAuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponse"
                }
              }
            }
          }
        },
        "summary": "Request a purpose-scoped OTP for a Member phone",
        "tags": [
          "Member Auth"
        ]
      }
    },
    "/api/v1/member/auth/otp/verify": {
      "post": {
        "operationId": "MemberAuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberSessionResponse"
                }
              }
            }
          }
        },
        "summary": "Verify an OTP and establish a Member session",
        "tags": [
          "Member Auth"
        ]
      }
    },
...
    "/api/v1/member/auth/revoke-all": {
      "post": {
        "operationId": "MemberAuthController_revokeAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberRevokedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke every active session for the Member",
        "tags": [
          "Member Auth"
        ]
      }
    },
    "/api/v1/member/auth/me": {
      "get": {
        "operationId": "MemberAuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberMeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Return server-authoritative Member identity",
        "tags": [
          "Member Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lottify v1 API",
    "description": "Authoritative REST/OpenAPI contract for Lottify v1",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AdminRevokedResponse": {
        "type": "object",
        "properties": {
          "revoked": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "revoked"
        ]
      },
      "AdminMeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "ADMIN",
              "AUDITOR"
            ]
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "accounting-period.read",
                "accounting-period.create-custom",
                "accounting-period.submit",
                "accounting-period.approve",
                "accounting-period.cancel",
                "accounting-period.close",
                "lottery-configuration.read",
                "lottery-configuration.create",
                "lottery-configuration.submit",
                "lottery-configuration.approve",
                "lottery-draw.read",
                "lottery-draw.manage",
                "result.read",
                "result.manage",
                "settlement.read",
                "settlement.manage",
                "withdrawal.read",
                "withdrawal.review",
                "withdrawal.payout",
                "promotion.read",
                "promotion.manage",
                "promotion.approve",
                "reconciliation.read",
                "report.read",
                "audit.read",
                "approval.read"
              ]
            }
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "capabilities"
        ]
      },
...
      "OtpRequestBody": {
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "enum": [
              "LOGIN",
              "REGISTER"
            ]
          },
          "phone": {
            "type": "string",
            "example": "+66812345678"
          }
        },
        "required": [
          "purpose",
          "phone"
        ]
      },
      "OtpRequestResponse": {
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "enum": [
              "LOGIN",
              "REGISTER"
            ]
          },
          "deliveredTo": {
            "type": "string",
            "description": "Masked delivery target"
          },
          "retryAfterSeconds": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "purpose",
          "deliveredTo",
          "retryAfterSeconds"
        ]
      },
      "OtpVerifyBody": {
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "enum": [
              "LOGIN",
              "REGISTER"
            ]
          },
          "phone": {
            "type": "string",
            "example": "+66812345678"
          },
          "code": {
            "type": "string",
            "pattern": "^[0-9]{4,8}$",
            "example": "123456"
          },
          "deviceName": {
            "type": "string",
            "example": "My Phone"
          }
        },
        "required": [
          "purpose",
          "phone",
          "code"
        ]
      },
      "MemberSessionResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "memberId": {
            "type": "string"
          },
          "accountCreated": {
            "type": "boolean"
          },
          "deviceId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "accessToken",
          "memberId",
          "accountCreated",
          "deviceId"
        ]
      },
...
      "MemberRevokedResponse": {
        "type": "object",
        "properties": {
          "revoked": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "revoked"
        ]
      },
      "MemberMeResponse": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ]
          }
        },
        "required": [
          "memberId",
          "phone",
          "status"
        ]
      },
...
    }
  }
}